+++ pipeline.py
@@ -21,7 +21,8 @@ AVAILABLE_MODELS = [
     'rigid_wasabi',
     'blaring_foundry',
     'flat_iceberg',
-    'drab_volcano'
+    'drab_volcano',
+    'icy-resin'
 ]
 
 # Lazy load only the model we are going to use
@@ -48,7 +49,9 @@ def load_pipeline_model(model_dir : str, model_name : str, override_batch_size=N
     if model_name == 'drab_volcano':
         from src.models.drab_volcano_model import drab_volcano_model
         model = drab_volcano_model()
-    
+    if model_name == 'icy-resin':
+        from src.models.icy_resin_model import icy_resin_model
+        model = icy_resin_model()
     if not os.path.exists(os.path.join(model_dir, model._checkpoint)):
         msg = f'Could not load {model_name}. Model checkpoint file "{model._checkpoint}" not found in model directory "{model_dir}"'
         log.error(msg)