+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 
 ### Added
+- Adding icy-resin model.
 - ML based map unit detection model Yolo_Legends
 - Added submodule containing support code for Yolo_Legends
 - Added Yolo_Legends checkpoint file to docker container
@@ -14,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 - Changed pipeline to use ML model for legend detection instead of heursitic method.
 - Updated cmaas_utils to 0.1.11, need to update package to run.
+
 
 ## [0.4.3] - 2024-08-06